+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Dev
 * support external attachments in Cucumber HTML reporter ([#182](https://github.com/vitalets/playwright-bdd/issues/182))
 * support trace-viewer in Cucumber HTML reporter ([#182](https://github.com/vitalets/playwright-bdd/issues/182))
+* hide BDD fixtures from Playwright HTML report
 
 ## 7.4.2
 * fix decorator typings for TypeScript consumers with `module` = `commonjs` ([#222](https://github.com/vitalets/playwright-bdd/issues/222))